@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <AssemblyName>plang</AssemblyName>
@@ -10,6 +10,7 @@
     <FileVersion>$(Version)</FileVersion>
     <SignAssembly>False</SignAssembly>
 		<PublishSingleFile>false</PublishSingleFile>
+		<LangVersion>default</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
