@@ -16,6 +16,6 @@ namespace PLang.Modules
 		Task<(Instruction? Instruction, IBuilderError? BuilderError)> Build(GoalStep step, Type responseType, string? errorMessage = null, int errorCount = 0);
 		Task<(Instruction? Instruction, IBuilderError? BuilderError)> Build(GoalStep step);
 		LlmRequest GetLlmRequest(GoalStep step, Type responseType, string? errorMessage = null);
-		void InitBaseBuilder(string module, IPLangFileSystem fileSystem, ILlmServiceFactory llmService, ITypeHelper typeHelper, MemoryStack memoryStack, PLangAppContext context, VariableHelper variableHelper, ILogger logger);
+		void InitBaseBuilder(string module, IPLangFileSystem fileSystem, ILlmServiceFactory llmService, ITypeHelper typeHelper, MemoryStack memoryStack, PLangAppContext context, VariableHelper variableHelper, ILogger logger, string methodName);
 	}
 }
