@@ -1,25 +1,89 @@
-﻿using PLang.Attributes;
-using PLang.Errors;
-using PLang.Errors.Runtime;
-using PLang.Exceptions;
+﻿using PLang.Errors;
 using PLang.Services.OutputStream;
 using System.ComponentModel;
-using System.Net.Http;
+using System.ComponentModel.DataAnnotations;
 using System.Text.RegularExpressions;
+using PLang.Services.Channels;
 
 namespace PLang.Modules.OutputModule
 {
+	public class AskProperties
+	{
+		[Required]
+		public string Question { get; set; }
+		public MessageType Type { get; set; } = MessageType.UserAsk;
+		public int StatusCode { get; set; } = 200;
+		public string? RegexPattern { get; set; } = null;
+		public string? ErrorMessage { get; set; } = null;
+		[Description("Allows user to add custom arguments")]
+		public Dictionary<string, object>? Args { get; set; } = null;
+		[NonSerialized]
+		public bool HasError = false;
+	}
+
+
+	public class WriteProperties
+	{
+		public object Data { get; set; }
+		public MessageType MessageType { get; set; } = MessageType.SystemAudit;
+		public int StatusCode { get; set; } = 200;
+		public Dictionary<string, object>? Args = null;
+	}
+	
+	
 	[Description("Outputs and writes out, to the UI a text or a variable. In console, code can ask user and he gives response")]
 	public class Program : BaseProgram, IDisposable
 	{
 		private readonly IOutputStreamFactory outputStream;
 		private readonly IOutputSystemStreamFactory outputSystemStream;
+		private readonly ChannelManager channelManager;
 
-		public Program(IOutputStreamFactory outputStream, IOutputSystemStreamFactory outputSystemStream) : base()
+		public Program(IOutputStreamFactory outputStream, IOutputSystemStreamFactory outputSystemStream, ChannelManager channelManager) : base()
 		{
 			this.outputStream = outputStream;
 			this.outputSystemStream = outputSystemStream;
+			this.channelManager = channelManager;
 		}
+
+		public Task<IError?> Write(string location, WriteProperties writeProperties)
+		{
+			/*
+			 * return await channel.Write({content, type, statusCode, callBack})
+			 */
+			var output = channelManager.GetChannel(writeProperties.MessageType);
+			if (output.Error != null) return Task.FromResult(output.Error);
+			
+			var error = output.Channel!.Write(writeProperties);
+			return error;
+		}
+
+		
+		
+		
+		public async Task<(string? Answer, IError? Error)> Ask(AskProperties askProperties)
+		{
+			/*
+			 * return await channel.Ask({content, type, statusCode, callBack})
+			 */
+			
+			var input = channelManager.GetChannel(askProperties.Type);
+			if (input.Error != null) return (null, input.Error);
+			
+			var answer = await input.Channel!.Ask(askProperties);
+			if (answer == null) return (null, null);
+			
+			// escape any variable that user inputs
+			answer = answer.Replace("%", @"\%");
+			if (askProperties.RegexPattern != null && !Regex.IsMatch(answer, askProperties.RegexPattern))
+			{
+				askProperties.HasError = true;
+				return await Ask(askProperties);
+			}
+			return (answer, null);
+		}
+		
+		
+		/*
 		[Description("Send response to user and waits for answer. type can be text|warning|error|info|debug|trace. statusCode(like http status code) should be defined by user. regexPattern should contain start and end character if user input needs to match fully. errorMessage is message to user when answer does not match expected regexPattern, use good grammar and correct formatting.")]
 		public async Task<string> Ask(string text, string type = "text", int statusCode = 200, string? regexPattern = null, string? errorMessage = null)
 		{
@@ -38,16 +102,20 @@ namespace PLang.Modules.OutputModule
 			}
 			return result;
 		}
-
+*/
 		public void Dispose()
 		{
+			/*
 			var stream = outputStream.CreateHandler();
 			if (stream is IDisposable disposable)
 			{
 				disposable.Dispose();
-			}
+			}*/
 			
 		}
+		
+		
+		/*
 		[Description("Write to the system output. type can be text|warning|error|info|debug|trace. statusCode(like http status code) should be defined by user. type=error should have statusCode between 400-599, depending on text")]
 		public async Task<IError?> WriteToSystemOutput(object? content = null, bool writeToBuffer = false, string type = "text", int statusCode = 200)
 		{
@@ -91,7 +159,7 @@ namespace PLang.Modules.OutputModule
 				await outputStream.CreateHandler().Write(content, type, statusCode);
 			}
 			return null;
-		}
+		}*/
 
 	}
 }
