@@ -26,7 +26,7 @@ namespace PLang.Modules.CallGoalModule.Tests
 		
 
 			builder = new Builder();
-			builder.InitBaseBuilder("PLang.Modules.CallGoalModule", fileSystem, llmServiceFactory, typeHelper, memoryStack, context, variableHelper, logger);
+			builder.InitBaseBuilder("PLang.Modules.CallGoalModule", fileSystem, llmServiceFactory, typeHelper, memoryStack, context, variableHelper, logger, "");
 
 		}
 
@@ -36,7 +36,7 @@ namespace PLang.Modules.CallGoalModule.Tests
 			if (llmService == null) return;
 
 			builder = new Builder();
-			builder.InitBaseBuilder("PLang.Modules.CallGoalModule", fileSystem, llmServiceFactory, typeHelper, memoryStack, context, variableHelper, logger);
+			builder.InitBaseBuilder("PLang.Modules.CallGoalModule", fileSystem, llmServiceFactory, typeHelper, memoryStack, context, variableHelper, logger, "");
 		}
 		public GoalStep GetStep(string text)
 		{
